Modification of previous test cases, and adding more test cases
